Add method signature for response headers
